Initial refactor and create pull handlers
