+++ src/EventGridEmulator/Program.cs
@@ -38,6 +38,8 @@ builder.Services.AddSingleton<ISubscriberCancellationTokenRegistry, SubscriberCa
 builder.Services.AddSingleton<IEventGridEventHttpContextHandler, EventGridEventHttpContextHandler>();
 builder.Services.AddSingleton<ICloudEventHttpContextHandler, CloudEventHttpContextHandler>();
 builder.Services.AddSingleton<CompositeEventHttpContextHandler>();
+builder.Services.AddSingleton<PullEventHttpContextHandler>();
+builder.Services.AddSingleton<PullQueueHttpContextHandler>();
 builder.Services.AddSingleton(typeof(TopicSubscribers<>));
 builder.Services.AddHostedService<ApplicationLifetimeLoggingHostedService>();
 builder.Services.AddHostedService<PeriodicConfigurationReloadHostedService>();
@@ -45,139 +47,13 @@ builder.Services.AddHostedService<PeriodicConfigurationReloadHostedService>();
 var app = builder.Build();
 
 app.MapPost(CompositeEventHttpContextHandler.Route, CompositeEventHttpContextHandler.HandleAsync);
-
-// TODO Move somewhere else
-app.MapPost("topics/{topic}:publish", async (HttpContext context, [FromRoute] string topic, [FromServices] CompositeEventHttpContextHandler handler) =>
-{
-    await CompositeEventHttpContextHandler.HandleAsync(context, topic, handler);
-    return Results.Ok(new object());
-});
-// TODO Move somewhere else
-app.MapPost("topics/{topic}/eventsubscriptions/{subscription}:receive", async (string topic, string subscription, CancellationToken cancellationToken, [FromServices] TopicSubscribers<CloudEvent> events) =>
-{
-    var result = await events.GetEventAsync(topic, subscription, cancellationToken);
-    return Results.Ok(new { value = new[] { new { brokerProperties = new { deliveryCount = 1, lockToken = result.LockToken }, @event = result.Item } } });
-});
-// TODO Move somewhere else
-app.MapPost("topics/{topic}/eventsubscriptions/{subscription}:acknowledge", (string topic, string subscription, AcknowledgeData data, [FromServices] TopicSubscribers<CloudEvent> events) =>
-{
-    var succeededLockTokens = new List<string>();
-    var failedLockTokens = new List<FailedLockToken>();
-    if (data?.LockTokens is not null)
-    {
-        foreach (var token in data.LockTokens)
-        {
-            if (token is null)
-            {
-                continue;
-            }
-
-            if (events.TryDeleteEvent(topic, subscription, token))
-            {
-                succeededLockTokens.Add(token);
-            }
-            else
-            {
-                failedLockTokens.Add(new() { LockToken = token, Error = new() { Message = "invalid token" } });
-            }
-        }
-    }
-
-    return Results.Ok(new
-    {
-        failedLockTokens,
-        succeededLockTokens,
-    });
-});
-// TODO Move somewhere else
-app.MapPost("topics/{topic}/eventsubscriptions/{subscription}:release", (string topic, string subscription, AcknowledgeData data, [FromServices] TopicSubscribers<CloudEvent> events) =>
-{
-    var succeededLockTokens = new List<string>();
-    var failedLockTokens = new List<FailedLockToken>();
-    if (data?.LockTokens is not null)
-    {
-        foreach (var token in data.LockTokens)
-        {
-            if (token is null)
-            {
-                continue;
-            }
-
-            if (events.TryReleaseEvent(topic, subscription, token))
-            {
-                succeededLockTokens.Add(token);
-            }
-            else
-            {
-                failedLockTokens.Add(new() { LockToken = token, Error = new() { Message = "invalid token" } });
-            }
-        }
-    }
-
-    return Results.Ok(new
-    {
-        failedLockTokens,
-        succeededLockTokens,
-    });
-});
-// TODO Move somewhere else
-// We don't support moving a message to the DLQ, so the logic is similar to acknowledge
-app.MapPost("topics/{topic}/eventsubscriptions/{subscription}:reject", (string topic, string subscription, AcknowledgeData data, [FromServices] TopicSubscribers<CloudEvent> events) =>
-{
-    var succeededLockTokens = new List<string>();
-    var failedLockTokens = new List<FailedLockToken>();
-    if (data?.LockTokens is not null)
-    {
-        foreach (var token in data.LockTokens)
-        {
-            if (token is null)
-            {
-                continue;
-            }
-
-            if (events.TryDeleteEvent(topic, subscription, token))
-            {
-                succeededLockTokens.Add(token);
-            }
-            else
-            {
-                failedLockTokens.Add(new() { LockToken = token, Error = new() { Message = "invalid token" } });
-            }
-        }
-    }
-
-    return Results.Ok(new
-    {
-        failedLockTokens,
-        succeededLockTokens,
-    });
-});
+app.MapPost(PullEventHttpContextHandler.PublishRoute, PullEventHttpContextHandler.HandlePublishAsync);
+app.MapPost(PullQueueHttpContextHandler.ReceiveRoute, PullQueueHttpContextHandler.HandleReceiveAsync);
+app.MapPost(PullQueueHttpContextHandler.AcknowledgeRoute, PullQueueHttpContextHandler.HandleAcknowledgeAsync);
+app.MapPost(PullQueueHttpContextHandler.ReleaseRoute, PullQueueHttpContextHandler.HandleReleaseAsync);
+app.MapPost(PullQueueHttpContextHandler.RejectRoute, PullQueueHttpContextHandler.HandleRejectAsync);
 
 app.Run();
-
-// TODO Move somewhere else
-internal sealed class AcknowledgeData
-{
-    public string?[]? LockTokens { get; set; }
-}
-
-// TODO Move somewhere else
-internal sealed class FailedLockToken
-{
-    public string? LockToken { get; set; }
-
-    public ResponseError? Error { get; set; }
-}
-
-// TODO Move somewhere else
-internal sealed class ResponseError
-{
-    [JsonPropertyName("code")]
-    public string? Code { get; set; }
-
-    [JsonPropertyName("message")]
-    public string? Message { get; set; }
-}
 
 // For integration testing purposes only in order to use WebApplicationFactory<TProgram>
 public abstract partial class Program
